+++ pymuonsuite/quantum/vibrational/phonons.py
@@ -1,7 +1,4 @@
-"""Phonons extracted from CASTEP results
-
-Requites casteppy installed
-
+"""
 Author: Adam Laverack and Simone Sturniolo
 """
 
@@ -17,19 +14,10 @@ import shutil
 import numpy as np
 import scipy.constants as cnst
 from ase import Atoms
-from ase import io as ase_io
 from ase.calculators.dftb import Dftb
 from ase.dft import kpoints
 from ase.phonons import Phonons
-from soprano.collection import AtomsCollection
-from soprano.collection.generate import linspaceGen
-from soprano.utils import seedname
 
-from pymuonsuite.io.castep import parse_final_energy, parse_castep_muon
-from pymuonsuite.io.magres import parse_hyperfine_magres
-from pymuonsuite.io.output import write_tensors
-from pymuonsuite.quantum.grid import calc_wavefunction, weighted_tens_avg
-from pymuonsuite.utils import find_ipso_hydrogen, create_displaced_cells
 try:
     from casteppy.data.phonon import PhononData
 except ImportError:
@@ -145,126 +133,3 @@ def get_major_emodes(evecs, i):
     major_evecs = np.real(major_evecs)
 
     return major_evecs_i, major_evecs, major_evecs_ortho
-
-def phonon_hfcc(cell_f, mu_sym, grid_n, calc='castep', pname=None,
-                ignore_ipsoH=False, solver=False, args_w=False,
-                ase_phonons=False, dftb_phonons=False):
-    """
-    Given a file containing phonon modes of a muoniated molecule, either write
-    out a set of structure files with the muon progressively displaced in
-    grid_n increments along the axes of the phonon modes, or read in hyperfine
-    coupling values from a set of .magres files with such a set of muon
-    displacements and average them to give an estimate of the actual hfcc
-    accounting for nuclear quantum effects.
-
-    | Args:
-    |   cell_f (str): Path to structure file (e.g. .cell file for CASTEP)
-    |   mu_sym (str): Symbol used to represent muon in structure file
-    |   grid_n (int): Number of increments to make along each phonon axis
-    |   calc (str): Calculator used (e.g. CASTEP)
-    |   pname (str): Path of param file which will be copied into folders
-    |       along with displaced cell files for convenience
-    |   ignore_ipsoH (bool): If true, ignore ipso hydrogen calculations
-    |   solver (bool): If true, use qlab to numerically solve the schroedinger
-    |       equation
-    |   args_w (bool): Write files if true, parse if false
-    |   ase_phonons(bool): If True, use ASE to calculate phonon modes. ASE will
-    |       use the calculator of the input cell, e.g. CASTEP for .cell files. Set
-    |       dftb_phonons to True in order to use dftb+ as the calculator instead.
-    |       If False, will read in CASTEP phonons.
-    |   dftb_phonons(bool): Use dftb+ to calculate phonons if true. Requires
-    |       ase_phonons set to true
-    |
-    | Returns: Nothing
-    """
-    cell = ase_io.read(cell_f)
-    sname = seedname(cell_f)
-    #Parse muon data using appropriate parser for calculator
-    if (calc.strip().lower() in 'castep'):
-        mu_index, ipso_H_index, mu_mass = parse_castep_muon(cell, mu_sym,
-                                                            ignore_ipsoH)
-    else:
-        raise RuntimeError("Invalid calculator entered ('{0}').".format(calc))
-
-    if ase_phonons:
-        #Calculate phonons using ASE
-        masses = cell.get_masses()
-        masses[-1] = mu_mass/cnst.u
-        cell.set_masses(masses)
-        evals, evecs = ase_phonon_calc(cell, dftb_phonons)
-    else:
-        # Parse CASTEP phonon data into casteppy object
-        pd = PhononData(sname)
-        # Convert frequencies back to cm-1
-        pd.convert_e_units('1/cm')
-        # Get phonon frequencies+modes
-        evals = pd.freqs
-        evecs = pd.eigenvecs
-
-    # Get muon phonon modes
-    mu_evecs_index, mu_evecs, mu_evecs_ortho = get_major_emodes(evecs[0], mu_index)
-    # Get muon phonon frequencies and convert to radians/second
-    mu_evals = np.array(evals[0][mu_evecs_index]*1e2*cnst.c*np.pi*2)
-    # Displacement in Angstrom
-    R = np.sqrt(cnst.hbar/(mu_evals*mu_mass))*1e10
-
-
-    # Write cells with displaced muon
-    if args_w:
-        for i, Ri in enumerate(R):
-            cell.info['name'] = sname + '_' + str(i+1)
-            dirname = '{0}_{1}'.format(sname, i+1)
-            lg = create_displaced_cells(cell, mu_index, grid_n, 3*mu_evecs[i]*Ri)
-            collection = AtomsCollection(lg)
-            for atom in collection:
-                atom.set_calculator(cell.calc)
-            collection.save_tree(dirname, "cell")
-            #Copy parameter file if specified
-            if pname:
-                for j in range(grid_n):
-                    shutil.copy(pname, os.path.join(dirname,
-                         '{0}_{1}_{2}/{0}_{1}_{2}.param'.format(sname, i+1, j)))
-
-    else:
-        # Parse hyperfine values from .magres files and energy from .castep
-        # files
-        E_table = []
-        hfine_table = ipso_hfine_table = np.zeros((np.size(R), grid_n))
-        num_species = np.size(cell.get_array('castep_custom_species'))
-        all_hfine_tensors = np.zeros((num_species, np.size(R), grid_n, 3, 3))
-        for i, Ri in enumerate(R):
-            E_table.append([])
-            dirname = '{0}_{1}'.format(sname, i+1)
-            for j in range(grid_n):
-                mfile = os.path.join(dirname,
-                    '{0}_{1}_{2}/{0}_{1}_{2}.magres'.format(sname, i+1, j))
-                mgr = parse_hyperfine_magres(mfile)
-                hfine_table[i][j] = np.trace(
-                    mgr.get_array('hyperfine')[mu_index])/3.0
-                if not ignore_ipsoH:
-                    ipso_hfine_table[i][j] = np.trace(
-                        mgr.get_array('hyperfine')[ipso_H_index])/3.0
-                else:
-                    ipso_hfine_table = None
-                for k, tensor in enumerate(mgr.get_array('hyperfine')):
-                    all_hfine_tensors[k][i][j][:][:] = tensor
-                castf = os.path.join(dirname,
-                    '{0}_{1}_{2}/{0}_{1}_{2}.castep'.format(sname, i+1, j))
-                E_table[-1].append(parse_final_energy(castf))
-
-        E_table = np.array(E_table)
-        if (hfine_table.shape != (3, grid_n) or
-                E_table.shape != (3, grid_n)):
-            raise RuntimeError("Incomplete or absent magres or castep data")
-
-        symbols = cell.get_array('castep_custom_species')
-
-        r2psi2 = calc_wavefunction(R, grid_n, E_table = E_table,
-            write_table = True, value_table = hfine_table, sname = sname)
-
-        hfine_tens_avg = weighted_tens_avg(all_hfine_tensors, r2psi2)
-        write_tensors(hfine_tens_avg, sname, symbols)
-
-        calc_harm_potential(R, grid_n, mu_mass, mu_evals, E_table, sname)
-
-    return
